// frontend/libs/aas-designer-core/src/lib/v3-editor/model/editor-type-option.ts
export enum EditorTypeOption {
  AllSubmodels = 'AllSubmodels',
  Submodel = 'Submodel',
  ConceptDescriptions = 'ConceptDescriptions',
  ConceptDescription = 'ConceptDescription',
  SupplementalFiles = 'SupplementalFiles',
  MissingSubmodel = 'MissingSubmodel',
  Assets = 'Assets',
  Asset = 'Asset',
  SupplementalFile = 'SupplementalFile',
  Property = 'Property',
  MultiLanguageProperty = 'MultiLanguageProperty',
  SubmodelElementCollection = 'SubmodelElementCollection',
  SubmodelElementList = 'SubmodelElementList',
  Operation = 'Operation',
  Range = 'Range',
  BasicEventElement = 'BasicEventElement',
  Event = 'Event',
  Entity = 'Entity',
  Blob = 'Blob',
  File = 'File',
  ReferenceElement = 'ReferenceElement',
  SubmodelElement = 'SubmodelElement',
  Capability = 'Capability',
  RelationshipElement = 'RelationshipElement',

  // check
  AssetAdministrationShells = 'AssetAdministrationShells',
  AssetAdministrationShell = 'AssetAdministrationShell',
  SnippetGroup = 'SnippetGroup',
  ElementGroup = 'ElementGroup',
  ContactInformation = 'ContactInformation',
  AddGroup = 'AddGroup',
  OptionalElements = 'OptionalElements',
  FromJson = 'FromJson',

  Undefined = 'Undefined',
  AasFileContent = 'AasFileContent',
}